+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Get rid of implementation args as they are not used.
 - Improve ripgrep tool status output.
 - Added ask_human tool to allow human operator to answer questions asked by the agent.
+- Handle keyboard interrupt (ctrl-c.)
 
 ## [0.6.4] - 2024-12-19