+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.7.0] - 2024-12-20
 
 - Make delete_tasks tool available to planning agent.
 - Get rid of implementation args as they are not used.
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Handle keyboard interrupt (ctrl-c.)
 - Disable PAGERs for shell commands so agent can work autonomously.
 - Reduce model temperatures to 0.
+- Update dependencies.
 
 ## [0.6.4] - 2024-12-19